Реализует корректный показ аватара пользователя со ссылкой на MyList на всех страницах
