+++ src/components/user-block/user-block.jsx
@@ -1,26 +1,39 @@
 import React from "react";
 import PropTypes from "prop-types";
+import {connect} from "react-redux";
 import {Link} from "react-router-dom";
+
+import {isAuthorized} from "../../utils";
+import {AppRoute} from "../../const";
 
 const avatarDefault = `img/avatar.jpg`;
 
-const UserBlock = ({isAuthorized, avatarUrl = avatarDefault}) => {
+const UserBlock = ({authorizationStatus, userAvatarUrl = avatarDefault}) => {
+  const isUserAuthorized = isAuthorized(authorizationStatus);
   return (
     <div className="user-block">
       {
-        isAuthorized
+        isUserAuthorized
           ? <div className="user-block__avatar">
-            <img src={avatarUrl} alt="User avatar" width="63" height="63" />
+            <Link to={AppRoute.MY_LIST}>
+              <img src={userAvatarUrl} alt="User avatar" width="63" height="63" />
+            </Link>
           </div>
-          : <Link to="/login" className="user-block__link">Sign in</Link>
+          : <Link to={AppRoute.LOGIN} className="user-block__link">Sign in</Link>
       }
     </div>
   );
 };
 
 UserBlock.propTypes = {
-  isAuthorized: PropTypes.bool.isRequired,
-  avatarUrl: PropTypes.string,
+  authorizationStatus: PropTypes.string.isRequired,
+  userAvatarUrl: PropTypes.string,
 };
 
-export default UserBlock;
+const mapStateToProps = ({USER}) => ({
+  authorizationStatus: USER.authorizationStatus,
+  userAvatarUrl: USER.authorizationInfo.avatar_url,
+});
+
+export {UserBlock};
+export default connect(mapStateToProps, null)(UserBlock);
